[FIX] get_coll_type behavior when missing coll_key
previously, behavior was inconsistent between coulomb & non-coulomb collisions. Also, it was crashing when obj.match_type == MATCH_AUX and no coll keys were found for coulomb collisions, which is wrong; in that case it should (and now does) return None (to indicate "use 0 for collisions between these species").

The full documentation of the new behavior can be found in the docstring of the function.

Note: this may cause crashes in places where the code was previously silently using 0 for collision frequency (due to missing coll key) in MATCH_PHYSICS mode for non-coulomb collisions.