Remove bloated dependencies
`xarray[io]` adds a lot of bloat and unnecessary packages when running headless. This package should be able to be installed without extensive plotting ability, so please add the xarray io as option when necessary. 

Likewise, radynpy should be an optional dependency. It cannot be installed via conda/mamba and is unnecessary for most people.